+++ src/main/java/by/makei/spring/dto/CompanyReadDto.java
@@ -1,4 +1,4 @@
 package by.makei.spring.dto;
 
-public record CompanyReadDto(Long id) {
+public record CompanyReadDto(Integer id) {
 }
