+++ src/main/java/by/makei/spring/service/CompanyService.java
@@ -1,10 +1,10 @@
 package by.makei.spring.service;
 
 import by.makei.spring.database.entity.Company;
-import by.makei.spring.dto.CompanyReadDto;
 import by.makei.spring.database.repository.CrudRepository;
 import by.makei.spring.listener.entity.AccessType;
 import by.makei.spring.listener.entity.EntityEvent;
+import by.makei.spring.dto.CompanyReadDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
@@ -15,11 +15,11 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class CompanyService {
 
-    private final CrudRepository<Long, Company> companyRepository;
+    private final CrudRepository<Integer, Company> companyRepository;
     private final UserService userService;
     private final ApplicationEventPublisher eventPublisher;
 
-    public Optional<CompanyReadDto> findById(Long id) {
+    public Optional<CompanyReadDto> findById(Integer id) {
         return companyRepository.findById(id)
             .map(entity -> {
                 eventPublisher.publishEvent(new EntityEvent(entity, AccessType.READ));
