+++ src/main/java/by/makei/spring/service/UserService.java
@@ -1,8 +1,8 @@
 package by.makei.spring.service;
 
 import by.makei.spring.database.entity.Company;
-import by.makei.spring.database.repository.UserRepository;
 import by.makei.spring.database.repository.CrudRepository;
+import by.makei.spring.database.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -11,5 +11,5 @@ import org.springframework.stereotype.Service;
 public class UserService {
 
     private final UserRepository userRepository;
-    private final CrudRepository<Long, Company> companyRepository;
+    private final CrudRepository<Integer, Company> companyRepository;
 }
